refactor(GUI): move SupportedFormatsModel to lib/shared
