+++ lib/gui/os/dialog/dialog.js
@@ -25,9 +25,7 @@
 
 const angular = require('angular');
 const MODULE_NAME = 'Etcher.OS.Dialog';
-const OSDialog = angular.module(MODULE_NAME, [
-  require('../../models/supported-formats')
-]);
+const OSDialog = angular.module(MODULE_NAME, []);
 
 OSDialog.service('OSDialogService', require('./services/dialog'));
