+++ lib/gui/os/dialog/services/dialog.js
@@ -19,8 +19,9 @@
 const _ = require('lodash');
 const electron = require('electron');
 const errors = require('../../../../shared/errors');
+const supportedFormats = require('../../../../shared/supported-formats');
 
-module.exports = function($q, SupportedFormatsModel) {
+module.exports = function($q) {
 
   /**
    * @summary Current renderer BrowserWindow instance
@@ -64,7 +65,7 @@ module.exports = function($q, SupportedFormatsModel) {
         filters: [
           {
             name: 'OS Images',
-            extensions: _.sortBy(SupportedFormatsModel.getAllExtensions())
+            extensions: _.sortBy(supportedFormats.getAllExtensions())
           }
         ]
       }, (files) => {
