+++ lib/gui/pages/main/controllers/image-selection.js
@@ -22,9 +22,9 @@ const path = require('path');
 const messages = require('../../../../shared/messages');
 const errors = require('../../../../shared/errors');
 const imageStream = require('../../../../image-stream');
+const supportedFormats = require('../../../../shared/supported-formats');
 
 module.exports = function(
-  SupportedFormatsModel,
   SelectionStateModel,
   AnalyticsService,
   ErrorService,
@@ -42,7 +42,7 @@ module.exports = function(
     'img',
     'iso',
     'zip'
-  ], SupportedFormatsModel.getAllExtensions());
+  ], supportedFormats.getAllExtensions());
 
   /**
    * @summary Extra supported extensions
@@ -51,7 +51,7 @@ module.exports = function(
    * @public
    */
   this.extraSupportedExtensions = _.difference(
-    SupportedFormatsModel.getAllExtensions(),
+    supportedFormats.getAllExtensions(),
     this.mainSupportedExtensions
   ).sort();
 
@@ -67,7 +67,7 @@ module.exports = function(
    *   .then(ImageSelectionController.selectImage);
    */
   this.selectImage = (image) => {
-    if (!SupportedFormatsModel.isSupportedImage(image.path)) {
+    if (!supportedFormats.isSupportedImage(image.path)) {
       const invalidImageError = errors.createUserError('Invalid image', messages.error.invalidImage({
         image
       }));
@@ -78,7 +78,7 @@ module.exports = function(
     }
 
     Bluebird.try(() => {
-      if (!SupportedFormatsModel.looksLikeWindowsImage(image.path)) {
+      if (!supportedFormats.looksLikeWindowsImage(image.path)) {
         return false;
       }
