+++ lib/gui/pages/main/main.js
@@ -50,7 +50,6 @@ const MainPage = angular.module(MODULE_NAME, [
   require('../../modules/analytics'),
   require('../../modules/error'),
   require('../../models/selection-state'),
-  require('../../models/supported-formats'),
   require('../../models/drives'),
 
   require('../../utils/byte-size/byte-size')
